fix review proposals
