+++ packages/client/src/api/Auth/index.ts
@@ -1,29 +1,5 @@
-import BaseApi from '../Base'
-
-export interface SigninData {
-  login: string
-  password: string
-}
-
-export interface SignupData {
-  first_name: string
-  second_name: string
-  login: string
-  email: string
-  password: string
-  phone: string
-}
-
-export interface UserInfo {
-  id: number
-  first_name: string
-  second_name: string
-  display_name: string
-  login: string
-  email: string
-  phone: string
-  avatar: string
-}
+import { BaseApi } from '@/api'
+import { SigninData, SignupData, UserInfo } from './types'
 
 class AuthApi extends BaseApi {
   constructor() {
