+++ packages/client/src/api/Auth/types.ts
@@ -1,0 +1,24 @@
+export interface SigninData {
+  login: string
+  password: string
+}
+
+export interface SignupData {
+  first_name: string
+  second_name: string
+  login: string
+  email: string
+  password: string
+  phone: string
+}
+
+export interface UserInfo {
+  id: number
+  first_name: string
+  second_name: string
+  display_name: string
+  login: string
+  email: string
+  phone: string
+  avatar: string
+}
