+++ packages/client/src/store/slices/index.ts
@@ -1,0 +1,1 @@
+export { default as authSlice } from './authSlice'
